# Correct Perf Overhead in benchmarks stemming from String allocation