--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,2 +1,5 @@
 [workspace]
 members = ["lr-core", "lr-derive", "lr-tests"]
+
+[profile.bench]
+lto = true