# Introduction
This PR corrects overhead that was identified in https://github.com/ncatelli/parcel-lr-benchmark which caused LR to perform significantly worse than it's equivalent parcel combinator implementation, which was counter-intuitive to what I expected. This looked to have been caused by an premature called to `format!` for an error String. Once removed performance drastically improved.

## Results
### Before (equivalent test)

```
simple_calculator_expr_parsing
                        time:   [21.560 µs 21.627 µs 21.710 µs]
                        change: [+1394.6% +1413.3% +1434.2%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 17 outliers among 100 measurements (17.00%)
  1 (1.00%) low severe
  3 (3.00%) high mild
  13 (13.00%) high severe
```

### After

```
simple calculator expression parsing/without tokenization
                        time:   [288.77 ns 290.18 ns 292.02 ns]
                        change: [-8.0385% -6.4372% -4.8349%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  2 (2.00%) high mild
  8 (8.00%) high severe

large expression/without tokenization
                        time:   [10.657 µs 10.678 µs 10.702 µs]
                        change: [-7.9292% -6.5964% -5.3594%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  1 (1.00%) low mild
  3 (3.00%) high mild
  10 (10.00%) high severe
```

## TODO
- [x] Run benchmarks in CI

# Linked Issues
resolves #21 
# Dependencies

# Test
- [x] Tested Locally
- [x] Documented

# Review
- [x] Ready for review
- [x] Ready to merge

# Deployment